+++ 42cursus/ft_printf/ft_putnbr_base.c
@@ -6,17 +6,17 @@
 /*   By: rgallego <rgallego@student.42madrid>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/22 19:15:13 by rgallego          #+#    #+#             */
-/*   Updated: 2021/10/26 11:05:09 by rgallego         ###   ########.fr       */
+/*   Updated: 2021/11/14 17:50:16 by rgallego         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
 int	ft_putnbr_base(unsigned long long nb, char *base, int len_base)
 {
 	if (nb < (unsigned long long)len_base)
-		return (write(1, &base[nb], 1));
+		return ((int)write(1, &base[nb], 1));
 	else
 		return (ft_putnbr_base(nb / len_base, base, len_base)
-			+ write(1, &base[nb % len_base], 1));
+			+ (int)write(1, &base[nb % len_base], 1));
 }
